+++ lib/unsure.dart
@@ -5,5 +5,5 @@ export 'src/ast.dart' show FormulaAst;
 export 'src/calculation.dart';
 export 'src/formatter.dart';
 export 'src/parser.dart';
-export 'src/range.dart';
+export 'src/normal_distribution.dart';
 export 'src/result.dart';